Move labware inside deck package (deck contains labware)
Restructure so src/deck/ is the top-level package with labware models
nested at src/deck/labware/. This reflects that the deck is the
higher-level concept containing labware items.

- src/labware/ -> src/deck/labware/ (model files unchanged)
- src/labware/deck/ -> src/deck/ (loader, schema, errors)
- All imports updated: src.labware -> src.deck

183 tests pass, no behavioral changes.
