+++ AGENTS.md
@@ -78,16 +78,17 @@ A modular system for executing experiment sequences defined in code or YAML.
 - **`camera.py`**: `Camera` class wrapping OpenCV for robust image capture (handles warmup, retries).
 - **`executor.py`**: `ProtocolExecutor` that orchestrates the `Mill` and `Camera` to run the compiled protocol.
 
-### Labware Models (`src/labware`)
-Geometry and positioning models for consumables on the deck.
+### Deck and Labware (`src/deck`)
+Deck configuration loading and labware geometry/positioning models.
 
-- **`labware.py`**: `Coordinate3D` and `Labware` base model. `Labware` now provides high-level shared behavior (e.g., `get_location`, `get_initial_position`) and common validation helpers; concrete required fields live in subclasses. All models use strict schema (`extra='forbid'`).
-- **`well_plate.py`**: `WellPlate(Labware)` for multi-well plates (e.g., SBS 96-well). Required fields include `name`, `model_name`, dimensions, layout (`rows`, `columns`), `wells`, and volume fields (`capacity_ul`, `working_volume_ul`). Also provides `get_well_center(well_id)`.
-- **`vial.py`**: `Vial(Labware)` for a single vial. Required fields include `name`, `model_name`, geometry (`height_mm`, `diameter_mm`), single `location`, and volume fields (`capacity_ul`, `working_volume_ul`), plus `get_vial_center()`.
+- **Labware models** (`src/deck/labware/`):
+  - **`labware.py`**: `Coordinate3D` and `Labware` base model. `Labware` now provides high-level shared behavior (e.g., `get_location`, `get_initial_position`) and common validation helpers; concrete required fields live in subclasses. All models use strict schema (`extra='forbid'`).
+  - **`well_plate.py`**: `WellPlate(Labware)` for multi-well plates (e.g., SBS 96-well). Required fields include `name`, `model_name`, dimensions, layout (`rows`, `columns`), `wells`, and volume fields (`capacity_ul`, `working_volume_ul`). Also provides `get_well_center(well_id)`.
+  - **`vial.py`**: `Vial(Labware)` for a single vial. Required fields include `name`, `model_name`, geometry (`height_mm`, `diameter_mm`), single `location`, and volume fields (`capacity_ul`, `working_volume_ul`), plus `get_vial_center()`.
 - **Deck configuration (YAML)**: Deck layout is defined in a **deck YAML** file (labware only; no gantry settings). Strict schema: only allowed fields; missing, extra, or wrong-type fields raise `ValidationError`.
-  - **`src/labware/deck/yaml_schema.py`**: Pydantic models for deck YAML: `DeckYamlSchema` (root, single key `labware`), `WellPlateYamlEntry` (two-point calibration points under `calibration.a1` and `calibration.a2`, axis-aligned only), `VialYamlEntry` (single vial location). Both require `model_name`. All use `extra='forbid'`.
-  - **`src/labware/deck/loader.py`**: `load_labware_from_deck_yaml(path)` loads a deck YAML file and returns `dict[str, Labware]` keyed by the names in the YAML. Well plates are built from calibration A1/A2 and x/y offsets (derived well positions); vials from a single explicit `location`.
-  - **`src/labware/deck/errors.py`**: `DeckLoaderError` for user-facing loader failures.
+  - **`src/deck/yaml_schema.py`**: Pydantic models for deck YAML: `DeckYamlSchema` (root, single key `labware`), `WellPlateYamlEntry` (two-point calibration points under `calibration.a1` and `calibration.a2`, axis-aligned only), `VialYamlEntry` (single vial location). Both require `model_name`. All use `extra='forbid'`.
+  - **`src/deck/loader.py`**: `load_labware_from_deck_yaml(path)` loads a deck YAML file and returns `dict[str, Labware]` keyed by the names in the YAML. Well plates are built from calibration A1/A2 and x/y offsets (derived well positions); vials from a single explicit `location`.
+  - **`src/deck/errors.py`**: `DeckLoaderError` for user-facing loader failures.
 - **Sample config**: `configs/deck.sample.yaml` — one well plate and one vial; use as reference for required fields and two-point calibration format.
 - **Sample inspection script**: `show_deck_objects.py` loads `configs/deck.sample.yaml` and prints the resulting object mapping.
 - **Usage**: Load labware with `load_labware_from_deck_yaml("configs/deck.sample.yaml")` to get e.g. `{"plate_1": WellPlate(...), "vial_1": Vial(...)}`. Resolve targets like `plate_1.get_well_center("A1")` for absolute XYZ.
