+++ progress/2026-02-12.md
@@ -559,3 +559,36 @@ Rename deck schema file and classes to YAML-specific names so future readers can
 ### Notes
 - Naming-only clean break; no behavioral changes.
 - Any external imports using old names will fail and must be updated immediately.
+
+---
+
+## Deck-contains-labware package restructure
+
+### Task
+Restructure so the `deck` package is the top-level module containing `labware` as a subpackage, instead of the previous `labware` containing `deck`.
+
+### Work completed
+- **Directory restructure**:
+  - `src/labware/` removed entirely
+  - `src/deck/` is now the top-level package
+  - `src/deck/labware/` contains labware models (`labware.py`, `well_plate.py`, `vial.py`)
+  - `src/deck/` contains deck-level modules (`loader.py`, `yaml_schema.py`, `errors.py`)
+- **Updated imports**:
+  - `src/deck/__init__.py`: exports all public symbols (labware models + loader + errors + schema)
+  - `src/deck/labware/__init__.py`: exports labware model symbols
+  - `src/deck/loader.py`: imports labware from `.labware` subpackage (was `..labware`)
+  - `tests/test_labware.py`: `from src.deck import ...` (was `src.labware`)
+  - `tests/test_deck_loader.py`: `from src.deck import ...` and `from src.deck.loader import ...` (was `src.labware` / `src.labware.deck.loader`)
+  - `show_deck_objects.py`: `from src.deck import ...` (was `src.labware`)
+- **Updated documentation**:
+  - `AGENTS.md`: updated section header and all file path references
+  - `README.md`: updated file path references and import example
+
+### Validation
+- Targeted: `pytest tests/test_deck_loader.py -q` and `pytest tests/test_labware.py -q`
+- Full suite: `pytest tests -q`
+- Sample script: `python show_deck_objects.py`
+
+### Notes
+- No behavioral changes; pure structural refactor.
+- All imports via `src.labware` are replaced with `src.deck`.
