+++ src/deck/__init__.py
@@ -1,8 +1,14 @@
 from .errors import DeckLoaderError
+from .labware import Coordinate3D, Labware, WellPlate, Vial, generate_wells_from_offsets
 from .loader import load_labware_from_deck_yaml, load_labware_from_deck_yaml_safe
 from .yaml_schema import DeckYamlSchema, VialYamlEntry, WellPlateYamlEntry
 
 __all__ = [
+    "Coordinate3D",
+    "Labware",
+    "WellPlate",
+    "Vial",
+    "generate_wells_from_offsets",
     "DeckLoaderError",
     "DeckYamlSchema",
     "WellPlateYamlEntry",
