+++ src/deck/errors.py
@@ -3,4 +3,3 @@
 
 class DeckLoaderError(Exception):
     """Human-friendly deck loader error intended for CLI output."""
-
