+++ src/deck/labware/__init__.py
@@ -1,8 +1,3 @@
-from .deck import (
-    DeckLoaderError,
-    load_labware_from_deck_yaml,
-    load_labware_from_deck_yaml_safe,
-)
 from .labware import Labware, Coordinate3D
 from .well_plate import WellPlate, generate_wells_from_offsets
 from .vial import Vial
@@ -13,8 +8,4 @@ __all__ = [
     "WellPlate",
     "Vial",
     "generate_wells_from_offsets",
-    "load_labware_from_deck_yaml",
-    "load_labware_from_deck_yaml_safe",
-    "DeckLoaderError",
 ]
-
